refactor: context와 contextProvider 분리
- context를 사용할 수 있게 export하는 과정이 길어지고, 컨텍스트 타입도
  사용하도록 했습니다. 기존에 서로 다르지만 연관 된 이 두 가지가 충분히
  짧아서 함께 두었는데, 이제는 분리 할 때가 된 듯하여 분리합니다.
