+++ src/index.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import HabitTracker from "HabitTracker";
-import { ModalProvider } from "contexts/ModalContext";
-import { RecordProvider } from "contexts/RecordContext";
+import { ModalProvider } from "contexts/ModalContextProvider";
+import { RecordProvider } from "contexts/RecordContextProvider";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
