+++ src/interface/context.d.ts
@@ -18,3 +18,7 @@ export type PeriodContextType = {
 export type PeriodHandleContextType =
   | ((direction: "prev" | "next") => void)
   | null;
+
+export type ModalContextType = boolean | null;
+
+export type ModalHandleContextType = (() => void) | null;
